requirements.txt
# Web 框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 資料庫
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 快取和訊息佇列
redis[hiredis]==5.0.1
celery==5.3.4

# 身份驗證
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 驗證和配置
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# 網頁爬蟲
playwright==1.40.0

# 限流
slowapi==0.1.9

# 監控
prometheus-client==0.19.0

# 測試
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 工具
python-dotenv==1.0.0
requests==2.31.0
